+++ app/src/main/java/edu/badpals/brunos/rounds/RegularRound.java
@@ -12,8 +12,8 @@ public class RegularRound implements Round{
     @Override
     public void boxerRoundScore(){
         String[] puntuaciones = this.roundScore.split(" - ");
-        this.redBoxerScore = (byte)Integer.parseInt(puntuaciones[0]);
-        this.blueBoxerScore = (byte)Integer.parseInt(puntuaciones[1]);
+        this.redBoxerScore = Byte.parseByte(puntuaciones[0]);
+        this.blueBoxerScore = Byte.parseByte(puntuaciones[1]);
     }
 
     @Override
